Order service

+++ Project2_Angular/src/app/admin-search-users/admin-search-users.component.ts
@@ -3,40 +3,53 @@ import { Router } from '@angular/router';
 import { Project2Service } from '../project2.service';
 import { User } from '../user';
 
+
 @Component({
   selector: 'app-admin-search-users',
   templateUrl: './admin-search-users.component.html',
   styleUrls: ['./admin-search-users.component.css']
 })
+
+
 export class AdminSearchUsersComponent implements OnInit {
 
-  session:string[];
+  // user:User;
+  session: string[];
   userList: User[];
-  constructor(private project2Service:Project2Service,private router:Router) {
+  constructor(private project2Service: Project2Service, private router: Router) {
 
-  //  
-   }
+    //  
+  }
 
   ngOnInit() {
-    this.project2Service.getSession().subscribe(data=>{
-      this.session=data;
-      if(this.session==null){
+    this.project2Service.getSession().subscribe(data => {
+      this.session = data;
+      if (this.session == null) {
         this.router.navigate(['/adminLogin']);
-      }else{
-        this.project2Service.getAllUsers().subscribe(data=>{
+      } else {
+        this.project2Service.getAllUsers().subscribe(data => {
           this.userList = data;
         });
       }
     });
   }
 
-  deleteByUsername(user:User){
-    console.log("delete: "+user.userName);
+  goEditUser(user: User) {
+   // userTransfer = user;
+    this.router.navigate(['adminHome/adminEditUser', { id: user.userName }]);
+  }
+  deleteByUsername(user: User) {
+    console.log("delete: " + user.userName);
+    if(confirm("Do you want to delete User: "+user.userName)) {
     this.project2Service.deleteByUsername(user).subscribe(
-      data=>this.project2Service.getAllUsers().subscribe(data=>{
+      data => this.project2Service.getAllUsers().subscribe(data => {
         this.userList = data;
       }));
+    }
   }
- 
+
+  
 
 }
+
+//export var userTransfer: User = new User();
